[FEATURE] Implement the text ingestor and EPUB3 strategy
The ingestor service is responsible for ingesting source documents
(EPUB2, EPUB3, Markdown, etc.) and creating application models that
Rails can work with. This commit adds the ingestion framework, as
well as the first ingestion strategy for EPUB3. Our expectation is
that as the project continues, we will write additional ingestion
strategies for other file formats.

The work included in this commit breaks down, roughly, as follows:

  - Adds initial, empty schema file
  - Builds out models and model associations
  - Adds paperclip gem and supporting config for storing resources
  - Adds text resource API routes, actions, and serializers
  - Adds EPUB3 ingestion to Manifold