[FEATURE] Configure the Manifold API for Docker containerization
This commit includes changes to gitignore, and the removal of the puma
pid file from the repo. It also adjusts the database configuration to
pick up the database host from env vars, which is needed for container
orchestration. Furthermore, it adds another puma configuration file
that's targetted at running in a docker environment, in which the app
is bound to a port rather than a unix socket.

This docker setup is very much a work in progress, and something we'll
want to continue to develop over time. For docker-compose.yml and the
dockerfile for this repo, head over to the manifold-docker repository:

https://github.com/ManifoldScholar/manifold-docker