[BUGFIX] Replace CSS ID selector with XPath ID selector
The EPUB3 strategy looks up EPUB3 rendition sources by ID. When this is
done using Nokogiri's CSS ID selector, the lookup will fail when there
are characters that are not allowed in CSS ID selectors (periods, for
example). The XPath selector is more forgiving, which is needed with some
of the sample EPUBs we've seen.