[CHORE] Include the GPL v3 license in the project
The decision to go with GPL was not undertaken lightly. The Manifold team discussed the merits of MIT vs GPL at length, and ultimately, we felt that the safeguards provided by the GNU license (particularly regarding the potential for patent abuse) are important ones. We hope to release some libraries and stand-alone aspects of Manifold separately, under the more permissive MIT licenses, but those decisions will be made as these libraries are developed. Ultimately, our hope is that any improvements to Manifold will be shared back with the larger community, and it is our belief that this is more likely to occur under the GPL license.