[CHORE] Rename DB and Manifold rake tasks
These tasks are still largely placeholder tasks, and we'll see if they
remain for long. That said, they were poorly named. This commit improves
the naming of the db:kill task, which kills postgres connections, and
moves the ingestion tasks out of the generic manifold namespace into a
more apt "ingest" namespace.