[FEATURE] Provide a manifold:install generator
We want developers to be able to easily install their own instance
of Manifold. With this commit, we're including a generator that can
be run with 'rails g manifold:install'. For now, all this does is
include config/secrets.yml. This generator will need to be updated
as Manifold requiers more configuration and installation options.