[CHORE] All repository code conforms to rubocop linting rules.
This large commit touches most files on the project. The changes
included are in response to the introduction of Rubocop with a
strict style ruleset. Note that there is also a minor ruby version
increase in this commit, to bring our Ruby version up to date with
the minor version whose syntax Rubocop is using.