[FEATURE] Add client application to the project.
This commit merges in a refactored version of what was initially the
manifold-reader proof of concept. The client is an isomorphic
node/react application built around Redux. With this commit, we're
adding the basic scaffolding.

Going forward, the API, a Rails app, will live in /api. The client, a
Node app, will live in /client.

The main entry point for the client is via NPM scripts. In development
start it up with npm run start-dev. This will start the webpack asset
server and the node client server. The node client server will proxy
the backend API so that you don't need to worry about CORS issues at
this point.

The client server is currently configured through a .env file in the
root. You can adjust the proxying, and the port it listens on by
modifying that file.