[F] Incorporate initial browse styles and containers
Advances work on PT #108400756. This commit begins the style
architecture for the browse views. It adds the basics for containers,
and also stubs files for STACSS (Structure Typography Appearance)
mixins that will be used to abstract common style patterns.

This commit also removes the normalizr dependency from the client. We
are now normalizing API data on the server-side, so there's no need to
re-normalize it on the client side. The API base client has been
adjusted to accomodate the new response style.