[FEATURE] Generate manifold icon font from svgs in dir
This commit unwinds the previous commit, replacing iconfont-loader with
fontgen-loader. It adds a proxy to the main application server that will
proxy webpack assets on 3001, which is needed to avoid CORS issues around
loading fonts. It also adds a proof of concept for an icon font, which
for now we're calling "manicons".