[F] Accept JWT based API auth; add user model
With this commit, the Manifold API has stateless token-based API
authentication. We're sticking to the JWT (JSON Web Token) spec.
To round things out, this commit adds a user model, a user controller,
and some improvements to the manifold install generator to create a
new admin user.

For more on JWT, see:
https://tools.ietf.org/html/rfc7519
http://jwt.io/