[F] Add thunk middleware to client store
Thunk middleware allows actions to return a function. The middleware
will call the function, passing it dispatch and getStore. The function
should, in turn, return an action object. Manifold needs this middleware
for the upcoming authenticated API fetch actions, which will need access
to state in order to grab the authToken and pass it in the request.