[F] Make client-side authentication possible
This commit implements client-side authentication for Manifold and
includes a number of changes:

1. The API interface has been refactored to rely on thunk-middleware.
Instead of returning a promise, API methods now return an object literal
that configured the API request. The action creators return a function,
which accepts dispatch and getState, passed to it by thunk middleware.
That function grabs the auth token from the store, injects it into the
API request configuration, and hands it off to the API client, which
returns a promise. That promise is then dispatched, only to later be
handled by the promise middleware.

2. Components have been refactored to use the connectData decorator. This
clears the way for restricting access to components using the "requireRole"
decorator, which is also included in this commit.

3. This commit also includes the login component, authentication action
creators, and authentication reducuers. Authentication related state is
now stored in state.authentication.

4. The server has been modified to look for an auth token cookie, which
contains the JWT. If it finds it, this token is injected into the store
on the server-side render.