[F] Add stub markup and styles for login overlay
[FEATURE #109484958] A stub file has been added with all the markup
necessary for the login overlay. Additionally, the login/user button
has been moved into its own component gets passed a boolean describing
whether the user is logged in or not. Stub methods have been added
based on this state for testing purposes.
Because the login overlay does not currently have an open/closed state,
it has been committed as closed by default, but it can be shown by
passing a parameter in Frontend.js