[F] Refactor login form to be a single container
[FEATURE #109484958] The login form has been moved to a single
container that is pulled in from both the login page (which is a top)
level route, and the login overlay (which is now a simple component).
Note that while the styles have not been set for the form on the
login page, component based CSS scoping allows both forms to be styled
separately while maintaining the same markup and functionality.