[F] Port log in overlay styles to login page
[FEATURE #109484958] The styles that control the structure of the
overlay log in form have been abstracted so that they may be shared with
the log in form when viewed on a page. Additionally, some specific
styles have been generalized. A log in page sass partial has been added
for primarily structural styles specific to that component. Initial
styles have also been built for the log out page UI.
NOTE: The verbiage for these buttons in the comps is always as
Log In (Two words) and the values in this commit have been changed to
reflect that. That said, this should be discussed and solidified early
on.