[B] Check for overlay before hiding on Log In
The log in container triggered a function that would hide a log in
overlay on a successful submission. This triggered an error when the
same form was included on a page without an overlay. Rather than check
both contexts for the existence of an overlay, an optional
"overlayVisible" property is set on the login form container
(default false) that the overlay closing action will run only if a true
value is passed in to the login form container.