[F] Add stub activity to project detail view
[FEATURE DELIVERS #108401236] Two components have been added to list
and render activity updates shown in the project detail view in the
comps. The list component accepts what is currently mock data from the
project detail container, and contains logic to match the heights in a
list of activity tiles.
The tile container itself contains several layout variants that will
be loaded depending on the "type" value of the update. Note that the
logic for determining and passing update types is entirely for markup
construction purposes at this time, and the data structures for these
activity updates are likely to change.
Currently this activity can be shown in its entirety or hidden based
on a minimum count declared in the project detail container.
The activity updates contain logic to center their content vertically
using padding calculated based on their parent list item height. Note
that this logic runs only once when the components load/are updated,
and it's not yet a final, responsive solution.