[F] Add stub component for project texts
[FEATURE DELIVERS #108401254] The static markup for the texts
component tries to approximate every permutation of a text thumbnail
output, but it is 10% static and not make any guesses about the data
structure from the API. When it is, this component will likely need
refactoring into smaller blocks of layout logic.
Like the hero, this is a initial responsive, but not mobile ready
layout.