[F] Add HTML validator and serializer
This commit adds two important pieces to the ingestion process, and
adjusts the TextSection model to store a json representation of its
HTML content.

The serializer is responsible for traversing a sections HTML content
and transforming it into a JSON object that includes all element nodes,
text nodes, and comment nodes. This structure can then in turn be
traversed by the client-side React application, which can use it to
create a shadow dom that matches the text content.

Initially, we thought we'd use dangerouslySetInnerHTML on the client-
side, but that approach is limited in that it prevents React from
knowing about the HTML content, and we can't leverage much of React's
one-way data flow. With this approach, we keep the unidirectional flux
architecture, and React's shadow DOM diffing mechanism even for large
HTML content.

Furthermore, this approach gives us the ability to decide what kinds of
content we won't support--eg, script tags--and remove those from the
source document as we serialize it.

There are some risks here. We need our ingested content to be well
formed, and the markup needs to play nice with React. To this end,
we're validating (and transforming) the HTML content prior to
serialization, which allows us to make small changes in order to make
the markup palatable to React.