[F] Add initial style and logic for reader TOC UI
[FEATURE #108401684] The markup and styles for the header have been
re-added from the stub-statics to work with new dynamic reader
components, and the reader-header title now contains the current text
title; Actions have been added for reader ui visibility, though it's
important to note that they have been cloned from frontend/ui actions
and may be absctracted/removed; The "contents" link uses one of these
to open and close the TOC drawer.
Components for the TOC drawer (and a more generic component for the TOC
 itself) have been built and receive the current text from Reader.js
for building the TOC as well as actions for closing the TOC drawer when
a link is clicked.