[F] User button has toggling dropdown menu
[FEATURE DELIVERS #110355096] In addition to some basic refactoring
around where share actions are stored for UI visibility, the user
button component has been built out into a menu component with a nested
button, and styled so that the menu can be shown/hidden with a Redux
action. It should be noted that this means the menu/button requires
some further refactoring to work in the browse view, as this work was
focused on the reader.
Basic logout functionality has also been added within the menu, but it
currently triggers an error that prevents the state from being
reflected in the front end. The above two catches will be addressed as
soon as possible.