[B] Improve logout behavior
When a user logs out of Manifold, what they see on the page may change,
as some of the content could be specific to their login (eg, whether or
not a project is followed). It's very difficult to clear only the user-
specific data out of the store's entity cache. A good long-term solution
could be to call all fetchData methods for currently visible components.
For now, however, we will just reload the current location after the
logout occurs.