[F] Add typography menu to reader
[FEATURE #110355246] The appearance panel has been built as static and
can be toggled by the user within the reader. A reducers have also been
built to increment/decrement font sizes and select the font style
(serif or sans-serif). All reducers past tests for function, and can
now be used in the front-end with actions. It's important to note that
these states do not persist to the user yet, but they use the same
initial data fields.