[F] Replace redux-react-router with redux-simple-router
Redux-react-router development is slowing down, and the community is
moving toward redux-simple-router, which aims to do less, and instead
push developers to rely more heavily on react-router's APIs. With this
revision, less information about the router is kept in the client's
store. Location, History, etc are passed down to components that are
tied to routes by react-router, and those components can pass them down
to child components as needed.

React-simple-router is not as compatible with the isomorphic react
starter kit foundation for the Manifold client as we would have liked.
Consequently, this commit is significantly revising the server-side
render, the client-side render, and the method for fetching data.

Before, we had transition middleware that would delay the transition
until all data for the route components had been fetched (via the
component fetchData methods). That worked well, but in some ways was
less than ideal because it delayed the transition by a few seconds.

With this revision, the server-side render relies on a fetchAllData
helper, whichi collects promises returned from component fetchData
methods and delays returning a response until they have all been
resolved.

The client-side render is a bit trickier. Here, we use a routing
context that wraps all components rendered by a router. This context
looks for components with a fetchData method and wraps them in a
component that delays the child component's rendering until fetchData
promises are resolved.

There is still some cleanup to be done, and ways we can reduce the
code here. This is being merged into the main branch a little bit
prematurely, and we intend to refine it in the short-term.