[F] Add static loading indicator component
[FEATURE DELIVERS #109480056] A global loading indicator component has
been added to both the browse and reader containers. This component is
currently a static element but has all of the markup/style necessary to
work as a global loading indicator bar.
This commit also corrects some minor linter errors in Reader.js

The loading indicator has three states:
Default:
Invoked on load
The indicator is invisible and cannot receive pointer events

Loading:
Invoked by adding ".loading" class to the loading-bar <div>
The indicator is visible and fills first quickly, and then increasingly
slowly as it approaches 90 percent full

Complete:
Invoked by adding ".complete" class to the loading-bar <div>
The indicator quickly fills to full from its previous state, while
simultaneously vanishing with a transition that completes half a second
after it is full

After completion, the bar currently needs to be switched to the default
state before loading/completion again. If this is even slightly
cumbersome to do in React, CSS animations can be employed
(instead of transitions) to reset the animation state for at the
"progress" stage.