[B] Featured project rows have bottom padding
[BUGFIX FIXES #111466604] Vertical padding is applied to grid rows
universally by applying padding to the grid's list-elements. In order
to prevent this from being applied to the bottom row, a negative margin
of the same value (as the row padding) is added to the wrapper. This
value is kept in a local variable for easy updating and scope safety.