[F] Add margin controls to reader
[FEATURE DELIVERS #111080632] New actions have been added to increment/
decrement margin values in the store, that use a now-abstracted version
of the font-size increment/decrement reducers. In the view, there are
now buttons to increase/decrease the margin size in the reader. It's
important to note that while these buttons do not have disabled states
for minimum and maximum sizes, they do stop changing the store when a
limit is reached.