[C] Configured scss linting for Hound CI
A configuration yml has been added to trigger a custom configuration of
scss-lint. This initial configuration is based on a combination of
established standards and also Cast Iron Coding house style. Notably
the style for variables, mixins, and functions all follow standard JS
convention (camelcase). We are also trying out recess CSS order.
This initial config is not final, and may be modified after a few
rounds of commits.