[C] Disables piping on server-side node app
We were using the piping module to reload the node server app when there
were changes to code in development. This was leading to a race condition
in which 1) a developer would change a file, 2) the change would trigger
the client-side update via HMR while 3) the server side app process was
being killed and restarted. As a result, the client-side HMR process would
not be able to perform hot module replacement because the server had gone
away. Piping and HMR seems relatively incompatible, so we're disbaling
piping for the time being.