[F] Add notification framework
Hurrah! Manifold can now notify its users when something goes wrong.

This commit adds actions for adding a notificatino (ADD_NOTIFICATION),
removing a notification (REMOVE_NOTIFICATION), and removing all
notifications (REMOVE_ALL_NOTIFICATIONS). The payload for adding a
notification takes this shape: { level, id, heading, body }

Level is an integer (0, 1, 2) that denotes the severity of the
notification. The ID is a unique identifier that can be used to remove
the notification in a subsequent action dispatch. Be aware that the ID
is an optional value in the payload; if it is not set, the action will
be automatically assigned a unique UUID (by the actionCreator).

The notification reducer will not allow two notifications that share
the same ID to exist in application state. If, for example, I add a
notification with the ID of "API_ERROR", then add another notification
with the same ID, the latter notification will replace the first in the
application store. This is by design, as it prevents the user from being
innundated with redundant notifications, as would be the case if the API
backend were not available and multiple calls failed.

The notification reducer also watches for any failed action (an dispatched
action with an error property of true). It then looks at the failed
action's payload, and determines whether it possesses the shape of a
normal notification: { level, id, heading body }. If it does, the
notification reducer will add that payload to its store, triggering the
rendering of a new notification. This feature is currently used to handle
rejected promises, mainly in the case of the API client. We'll see if this
is the right approach, and adjust as necessary.