[F] Show section title when user scrolls down
[FEATURE DELIVERS #112837675] In the reader, the header banner now
shows the title of the text-section after the user begins to scroll
down into the reader. The design logic behind this is that the user
almost always sees the section title when the reader is at the top of a
given section, so showing the text title gives the most context. Once
the user is further into the section, we switch to the section title to
provide a persistent reminder of the user's location in the TOC.

From a front-end standpoint, this section title is always in the markup
and the swap happens with a CSS translateX function.

In addition to this markup, a layout bug has also been addressed which
ensures that the titles are truncated if they are longer than their
allotted space.

It should be noted that the section title value is dependent on commit
1e26ef412c78182e6dc3a289f5cb62597de1c21e (submitted as a pull request
in feature/pagination) as it gets the sectionId which is added to the
reader props in that commit. If the pagination branch is rejected for
any reason, this will need to be substituted.