[F] Change text-fade on reader dark color scheme
[FEATURE DELIVERS #113763363] In addition to swapping banner gradient
colors for the reader when it is in "dark mode," the banner gradient
has also been refactored as a DOM element (instead of a pseudo element)
and removed from the browse/frontend interface. In its place, a thin
border is used for now, but it should be noted that this is awaiting
approval from design.