[F] Upgrade routers and add section scrolling
Another day, another react-router and react-router-redux upgrade to
contend with. With this commit, we bump these to the most recent
versions, which brings breaking changes and refactoring, also addressed
in this commit.

That said, the react-router 2 API feels cleaner, and this commit adds a
basic but functional POC implementation of scrolling to URL fragments
in the reader section views, which is an important step forward.

Alas, this upgrade introduces a regression in the isomorphic rendering,
which causes react to discard the server side render and output a
warning. This ought to be fixed soon in a subsequent commit.