[B] Remove scroll-behavior
The scroll-behavior module is not mature enough for use on this
project. One problem is that it will scroll to the topeven when only
the request hash has changed, which effectively breaks anchor links.
For now, I'm replacing it with simple scroll-to-top behavior for a
handful of containers, but we'll likely need a better solution at
some point.