[B] Address Paperclip spoofing false positives
On the staging host, Paperclip repeatedly reports non-spoofed
attachments as spoofed, which breaks ingestion. We're quick-fixing it
with a monkey patch for now, and in doing so will incur some technical
debt. We'll circle back later and attempt to come up with a more robust
fix, perhaps by avioding paperclip alltogether.

See https://github.com/thoughtbot/paperclip/issues/1429 for more on
this issue.