[B] Solve unitialized constant when running tests
In the test environment, config.cache_classes had the effect of
triggering an uninitialized constant exception if a model class was
changed between test runs. In theory, spring is setting cache_classes
to false automatically, but we found that it had to be set explicitly
to avoid errors.

See the following issues for context:
https://github.com/rails/spring/issues/384
https://github.com/rails/spring/issues/95