[B] Refactor container data fetching
Boy, this is a tricky part of Manifold. In a perfect world, the
following things would happen when a user makes a request to the
Manifold client application:

1. The server-side express app determines what containers are being
rendered based on the request, and calls the fetch data method for
each of them.

2. The data is fetched synchronously, and stored in the Redux state

3. The initial render is returned to the browser, at which point the
client application takes over.

4. The client application knows that the data was already fetched on
the server, so it does not re-fetch the data, nor does it delay the
rendering of the containers, which leads to an invalid client/server
checksum.

With this commit, we find what I hope will be a simple and elegant
solution to this problem. In a nutshell, we track the rendering state
in the isomoprhic section of the store. The progression through the
states is NOT_LOADED > SERVER_LOADED > CLIENT_LOADED.

This commit adjusts the ResolveDataDependencies higher order
component so that it will not delay container rendering or try to
fetch data when the state is SERVER_LOADED and data is being fetched
in a client execution context. This effectively prevents the client
application from loading data until it is fully rendered.