[B] Load DevTools in client only lifecycle method
We kept finding that loading DevTools required either a duplicate
render in the client (render, compare server-client checksums,
then render again with DevTools). A good work around seems to be to
load the DevTools only on the client, and via a component lifecycle
method that only runs on the client.