[F] Update Babel, RHL & improve webpack
This is a large commit with a number of changes, all of which work
together to accomplisht the following:

- Babel is upgraded from v5 to v6
- React is upgraded from 14 to 15
- ESLint is updated and configuration is made current
- React Hot Loader v2 is replaced by React Hot Loader v3, which is beta
but functional
- The Manifold client server no longer proxies the Webpack asset server,
which means that we can re-enable piping on the server
- Javascript decorators, currently a stage 1 proposal, are removed in
keeping with Babel's decision to remove them until they stabilize, and
with React Hot Loader's current lack of support
- Technical debt is paid down throughout the application. Static class
functions, used early on to avoid manual bindings, have been removed
in keeping our current approaches.
- Package.json has been cleaned up and compatibility with NPM v3 has
been confirmed.