[F] Transform CamelCase params to underscored
The Manifold client, in keeping with current JSON style, expects API
results to be camelCased. To that end, active model serializers in the
Manifold API has been configured to transform underscored rails
properties to camel cased keys. This same transformation needs to
happen when the client sends params to the API.

For rationale here, see:

https://google.github.io/styleguide/jsoncstyleguide.xml?showone=Property_Name_Format#Property_Name_Format