[F] Allow users to create and update accounts
This is the first UI component that sends a PUT or POST request to the
API. As such, it revealed the need for refactoring in multiple places.
This commit introduces a Global component that wraps Frontend, Reader,
and the future Admin containers. Login, metadata, and other bits and
pieces are promoted to this top level container.

This commit also revises the authentication reducer, and how we store
the current user in state.