[B] Make Me + Users APIs consistent with JSON API
Create and Update calls to the API should include params that are
structured similar to the API responses (eg: { data: { attributes: {}}}).
This approach requires some helper methods in our controller validation
concern, which was stubbed out in the favorites/following branch, and is
now being backported to the users/me endpoints.