[B] Bring auth components in line with current API
Prior to this commit, we only stored the user attributes in the auth
portion of the store. It's cleaner to store the entire data object,
which includes attributes, metadata, and relationships in the store.
This makes how authentication stores its data more consistent with the
broader entityStore. This commit also changes how data is posted to the
/me and /users API endpoints to be more in keeping with the JSON API
specification.