[C] Refactor common container actions
Prior to this commit, header actions were imported and bound separately
in the Frontend and Reader components. This commits adds a helper method
that makes it simpler to get common action creators as props and pass
them down to children. In doing so, it dries up our code a bit.

Pinging @naomiyaki on this, as it touches on her earlier work.