[B] Do not reload the page on logout
Handling logout is rather complex. When a logout action is triggered, we
can clear up the application state using a reducer. That's fine, but in
most cases a logout needs to re-fetch data from the API to ensure that
the data cached in state is current and reflects the user's
unauthenticated state. To accomplish this, we're connecting the higher
order ResolveDataDependencies component to the store and passing it
authentication.authenticated as a prop. When the prop changes from
logged in to logged out (or vice versa), ResolveDataDependencies will
re-render its children, which will cause it to call the fetchData
methods on those children. In theory, this will cause the UI to update
with new data appropriate for the unauthenticated (or authenticated)
user.

[Delivers #121527391]