[F] Manifold can persist and display annotations
This commit includes our early work toward persisting annotations,
retrieving them from the API, and rendering them within the text.
There are known performance problems with the current approach, which
we believe can be addressed by better dirty checking within components.
For now, we are pushing this into the development branch so it's easier
to gauage performance problems without the overhead of our development
tooling.