[F] Refactor how app is served in development env
This breaks apart our approach to serving the client in the interest of
increased performance in development and reduced build times. Running
npm start kicks off a number of distinct processes:

1. We build some vendor modules that change very infrequently to
/dist/build/universal/dll. This way, webpack doesn't have to watch
these files or worry about them when its building changed files.

2. The entire application is built to dist/build/server/react-server.
This javascript is responsible for the server-side rendering of
application routes.

3. Just enough javascript to render an HTML container for the client-
side code is built to dist/build/server/web-server. In the event that
the universal page rendering service is not available, the application
can still render the HTMl container and let the client-side code take
over.

4. The "react server" begins. This is the express app at defined at
src/server-react. It includes the react-server js from #2, above.

5. The "web server" begins. This is mostly a proxy server that proxies
requests to the react server, or to the asset server. In the event
that the react-server is unreachable, which is often the case when its
reloading due to code change, the webserver will return an HTML
container that the client-side code can mount on.

6. The "asset server" begins. This is the webpack dev server that
serves the client-side assets in development mode. It also serves
all other files in /dist.