[F] Remove resource association from ingestion src
Initially, each ingestion source belonged to a resource. Given how our
concept of resources has developed, it no longer makes sense to think of
the assets that make up an epub as a resource that will be displayed on
the frontend. To clear the way for the actual resource model, we're
removing the relationship between ingestion sources and resources, and
instead storing the backing file on the ingestion source itself, via an
attachment field.