[F] Adds creator association to models
Manifold needs to be able to track which user creates a given model. For
the time being, we'll track this on projects, texts, resources, and
pages. This commit also makes all belongs_to relationships required,
unless otherwise specified. This validation was previously set, but not
enforced, likely due to an incompatibility in a third party gem (likely
friendly ID). Requiring creators leads to needing to refactor the
ingestion mechanism, so that there's a creating user present, which in
turn means creating seed data that creates a default CLI user model.