[F] Create two request metas for one XHR request
This small but significant change to the entityStore middleware allows
for request meta that is an array. We key results in the
store.entityStore.requests by the request meta, which allows containers
to later look up the results of requests that they triggered. With this
change, it is possible to pass an array of meta strings. This allows a
single request to appear to components as two requests, which can
reduce the number of XHR requests to the API. For example, on the
resource collection detail view, we have a slide-show and a list of all
collection components below. These are initially populated with the same
data, which is the first page of resources for that collection. However,
because the two components operated independently of each other, we
want to decouple the requests. With this change, the first request can
feed both components.