[B] Correct form state connect bugs
- The change to Set.js allows the Set component to pass an empty string
to a form input as its value prop when the property is undefined in the
session dirty model. This is often the case, for example, when we're
creating a new record, and there are no attributes on the dirty model.
Without this change, the input has an undefined value, and React
complains that is is switching from an uncontrolled to controlled input
when the user starts typing.
- The change to Form.js corrects which props are being used when the
form updates. Prior to this revision, the form would look at previous
props after the component updated to see if it needed to open a session.
It seems to me that it ought to be looking at the current props in this
case.
- Finally, the change to the new category form sets the session key
correctly.