[B] Improve edit session change detection
With this change, the editSession reducer no longer clones the entire
model to `dirtyModel`. Instead, it keeps a read-only version of the
model being edited, and stores new values in the dirty object. If the
dirty value is set to the same value as the source model, the reducer
will simply unset the dirty value. This allows for better detection
of what's actually changed, and whether or not anything has changed.

[Delivers #136843609, #137096285]