[F] Load settings in client; expose two settings
This commit represents initial work on exposing application settings via an API
endpoint. The wrapper Manifold container fetches settings when it is mounted if
the settings have not yet been loaded. In theory, this will take place on the server-
side render, and the settings will be present in the store when the page is loaded.
That said, we may need to revisit how settings are loaded.

This commit also adds default publisher settings and makes them editable via the backend.

[Delivers #134609689]