[F] Render install statistics in backend; adds GA
This commit add new values to the settings singleton, and adds a new
statistics resource available at /api/v1/statistics. This resource
exposes information about the installation such as the number of
annotations in the last week. It is as of yet not fully fleshed out,
but the foundation is present. The statistics resource will also call
out to the Google Analytics API, if GA is configured, to report on
number of readers, etc.

[Delivers #136390675]