[F] Bootstrap data and track page views
This commit adds a mechanism to the server-side render that calls the
static bootstrap method on the Manifold component. That fetch is
blocking, similar to how fetchData works, and allows the server-side
render to ensure that app settings are loaded before the client-side
render takes place. In the event that the server-side render is
unavailable, the settings will be loaded client-side. This commit
also adds google analytics integration. Page views should be tracked
on load and on route changes. Currently, this is running in debug
mode when __DEVELOPMENT__ is true. This may be disabled after we’re
confident it is working as expected.