[F] Text node UUID are stable during ingestion
Manifold assigns unique IDs to text nodes in texts and uses these to
track the position of annotations within a text. Prior to this commit,
those IDs were randomly generated during text ingestion, which made
them inherently unstable. With this commit, the UUIDs are based on
the hash of the node contents. In cases of duplicate identifiers, the
sequence of the duplicate node is taken into consideration as well.
This is not a good long-term solution. If there are 15 text nodes in the
document with the content “ “, changing the first one will change all
subsequent node identifiers. In the future, the mechanism by which
these identifiers are assigned should take into consideration the
relative context of the node rather than it’s sequence within the larger
document.