[C] Refactor configuration
In manifold we have the concept of configuration—server-side, mostly concealed API configuration—and settings—client side, mostly exposed application settings. We also have secrets that are stored in the environment. These distinctions have made it difficult to know where to store things. With this commit, we present the API a unified interface—Rails.configuration.manifold—to access all API configuration. Secrets stored in the environment will be rolled up into this global API configuration hash (er, mash, via Hashie). With this commit, we stop accessing ENV vars directly in the API code.

This commit also works around an issue we’re seeing with multiline .env values and Ubuntu16, such as the google service private key. Now, keys can be stored in /config/keys, an untracked directory that is shared across deploys. If the google service private key is found there, it will be used in place of any environment variables.