[C] Refactor Form.Upload component
With this commit, the Form.Upload component accepts a “style” property
of “portrait,” “landscape,” or “square.” The component also takes a new
“accepts” string property. This property configures the underlying drop-
zone component accepted mime types. If the “accepts” property is set to
“images,” the component will render the current or pending  attachment
as a preview. Otherwise it will show a generic file icon and the current or
pending filename.